Add tpl.js file for template util에 포함되어 있던 템플릿 문자열들을 모아서 별도의 파일에 저장하였습니다.
